Support `raise_error()` on [databricks] 14.3, Spark 4. (#11969)
Fixes #10969.

This commit adds support for `raise_error()` on Databricks 14.3 and
Spark 4.0.

On these new Spark versions, the `RaiseError` expression (that powers
the `raise_error()` API function) was changed from a Unary expression to
a Binary one. This was done without modifying the arity of
`raise_error()`. The ostensible reason seems to have been to eventually
allow user-code to raise custom errors via `raise_error()`.

This commit allows `raise_error()` to work on the GPU as it currently
does on the CPU: as a unary function powered by a binary expression in
the background.

The tests have been modified to verify both the new behaviour and the
legacy one on new platforms, while continuing to run as before on legacy
platforms.

---------

Signed-off-by: MithunR <mithunr@nvidia.com>