Ensure Java Compatibility Check for Regex Patterns (#11912)
Resolves #11651

This PR updates the Parser.parse() function to include a call to
java.util.regex.Pattern.compile(pattern), ensuring that the pattern is
validated upfront. If the pattern is not valid in Java, the same error
should be raised for both CPU and GPU runs. This PR also modifies few Scala tests to align with the updated
functionality.

---------

Signed-off-by: Suraj Aralihalli <suraj.ara16@gmail.com>