HybridParquetScan: Refine filter push down to avoid double evaluation (#12000)
Closes https://github.com/NVIDIA/spark-rapids/issues/11892

In the current code, a HybridParquetScan followed by a Filter will
result in all conditions being pushed down to the CPU, but still
remaining in the Filter at the same time, so the Filter conditions are
evaluated twice. Usually the second evaluation is quite fast, so this
won't be a big problem. But if there are some conditions that are not
supported by CPU or GPU, it will cause some problems.

This PR adds a rule to check each condition in filterExec before
overriding:
- If a filter condition is not supported by either CPU or GPU, it will
fallback to CPU in FilterExec and not push down to CPU.
- If a filter condition is only supported by the CPU, this pr pushes it
down to the scan and removes it in FilterExec.
- If a filter condition is only supported by the GPU, this pr keeps it
in the filter.
- If all conditions are pushed down to the scan, FilterExec is removed.

The `supportedByHybridFilters` is from
[velox-backend-support-progress](https://github.com/apache/incubator-gluten/blob/c8284e53f2eac5e02be79d5e3914a9b44a9f6f90/docs/velox-backend-support-progress.md)
in Gluten. Here is the script to extract CPU supported exprs,
[gist](https://gist.github.com/thirtiseven/5d9c3da285b41b2d1c409289fcc01b87).

For example:
```
scala> val df = spark.read.parquet("parse_url_protocol")
df: org.apache.spark.sql.DataFrame = [url: string, pr: string ... 2 more fields]


scala> df.filter("startswith(pr, 'h') == False and ascii(url) >= 16").show()
25/01/23 16:32:31 WARN GpuOverrides: 
!Exec <CollectLimitExec> cannot run on GPU because the Exec CollectLimitExec has been disabled, and is disabled by default because Collect Limit replacement can be slower on the GPU, if huge number of rows in a batch it could help by limiting the number of rows transferred from GPU to CPU. Set spark.rapids.sql.exec.CollectLimitExec to true if you wish to enable it
  @Partitioning <SinglePartition$> could run on GPU
  *Exec <ProjectExec> will run on GPU
    *Expression <Alias> cast(count#2L as string) AS count#32 will run on GPU
      *Expression <Cast> cast(count#2L as string) will run on GPU
    *Exec <FilterExec> will run on GPU
      *Expression <Not> NOT StartsWith(pr#1, h) will run on GPU
        *Expression <StartsWith> StartsWith(pr#1, h) will run on GPU
      *Exec <FileSourceScanExec> will run on GPU
```
`startswith` is not supported in CPU so it will be kept in GPU, and
`ascii` pushed down to CPU. The check is recursive.
![Screenshot 2025-01-23 at 16 38
42](https://github.com/user-attachments/assets/2732625e-b552-4c06-b663-00dafa0d1962)


and
```
scala> df.filter("url >= 'http' and ascii(url) >= 16").show()
25/01/23 16:37:45 WARN GpuOverrides: 
!Exec <CollectLimitExec> cannot run on GPU because the Exec CollectLimitExec has been disabled, and is disabled by default because Collect Limit replacement can be slower on the GPU, if huge number of rows in a batch it could help by limiting the number of rows transferred from GPU to CPU. Set spark.rapids.sql.exec.CollectLimitExec to true if you wish to enable it
  @Partitioning <SinglePartition$> could run on GPU
  *Exec <ProjectExec> will run on GPU
    *Expression <Alias> cast(count#2L as string) AS count#66 will run on GPU
      *Expression <Cast> cast(count#2L as string) will run on GPU
    *Exec <FileSourceScanExec> will run on GPU
```
if all filters are supported, the FilterExec will be removed.
![Screenshot 2025-01-23 at 16 38
29](https://github.com/user-attachments/assets/b132b043-b96e-42d2-87f3-1fd61bef9e91)


<!--

Thank you for contributing to RAPIDS Accelerator for Apache Spark!

Here are some guidelines to help the review process go smoothly.

1. Please write a description in this text box of the changes that are
being
   made.

2. Please ensure that you have written units tests for the changes
made/features
   added.

3. If you are closing an issue please use one of the automatic closing
words as
noted here:
https://help.github.com/articles/closing-issues-using-keywords/

4. If your pull request is not ready for review but you want to make use
of the
continuous integration testing facilities please label it with `[WIP]`.

5. If your pull request is ready to be reviewed without requiring
additional
   work on top of it, then remove the `[WIP]` label (if present).

6. Once all work has been done and review has taken place please do not
add
features or make changes out of the scope of those requested by the
reviewer
(doing this just add delays as already reviewed code ends up having to
be
re-reviewed/it is hard to tell what is new etc!). Further, please avoid
rebasing your branch during the review process, as this causes the
context
of any comments made by reviewers to be lost. If conflicts occur during
review then they should be resolved by merging into the branch used for
   making the pull request.

Many thanks in advance for your cooperation!

-->

---------

Signed-off-by: Haoyang Li <haoyangl@nvidia.com>
Co-authored-by: Alfred Xu <lovedreamf@gmail.com>