Relax decimal metadata checks for mismatched precision/scale [databricks] (#12060)
This PR fixes https://github.com/NVIDIA/spark-rapids/issues/11433 .

This PR makes the GPU Parquet reader more flexible when handling files
whose decimal columns have a different precision/scale than Spark’s
requested schema. Previously, the plugin’s code would fail early
(“Parquet column cannot be converted”) if the file declared, for
example, DECIMAL(20, 0) but Spark asked for DECIMAL(10, 0) or DECIMAL(5,
1). Now we defer these mismatches to be resolved with optional half-up
rounding or overflow handling trying to match standard Spark behavior.

In this PR, we make castDecimalToDecimal as public function. In
`evolveSchemaCasts`, we pass the from and to DecimalTypes to cast the
decimals to the required form. castDecimalToDecimal will handle the case
of widening the scale/precisions or narrowing of scale/precisions.

Updated the current integration tests. 
In Spark-UT we are disabling the test as the Apache Spark vectorized
path throws error where as the spark-rapids implementation produces the
correct results.

---------

Signed-off-by: Niranjan Artal <nartal@nvidia.com>