Quick fix for hybrid tests without git information.
For some tests in CI/CD, we only need to download specific test files and do not need to clone the entire Git repository.

As a result, we cannot retrieve the project root directory using Git commands.

So we use the $WORKSPACE(always pointing to the root dir of the project) instead of the Git top-level path.

Signed-off-by: timl <timl@nvidia.com>