Support HiveHash in GPU partitioning (#12192)
(No issue for this)

This PR adds the support to allow GPU hash partitioning using the same
hash function as the CPU one by trying to infer the type of hash
function from the CPU hash partitioning.

This is desgined for some Spark distributions supporting to specify a
hash function (e.g. one of `HiveHash`, `Murmur3Hash` ...) for hash
partitioning by introducing a new field named "hashingFunctionClass". It
is not like the regular Spark who always use `Murmur3Hash`. However, to
align with the regular Spark's behavior, `Murmur3Hash` is still the
default option when inferring fails, and of course the inferrring will
always fail in regular Spark.

Now only `HiveHash` and `Murmur3Hash` are supported on GPU.

This is hard to test by unit tests or integration tests. Since the
customized Spark distribution is not public and it does not have any
side effects to the regular Spark. But the CI at least can ensure no
regressions and we already verify it by customer queries with the
customized Spark.

---------

Signed-off-by: Firestarman <firestarmanllc@gmail.com>