Materialize the stream side first for BroadcastNestedLoopJoins [databricks] (#12257)
Fixes #12256

The PR is about to pick up the missing trick for the implementation of
BroadcastNestedLoopJoin: materialize the stream side in front of the
broadcast build batch, which has been applied on the BroadcastHashJoin.

In this PR, the private method of BCHJ
`getBroadcastBuiltBatchAndStreamIter` is generalized as a common helper
method. And it is applied onto BCNLJ as well.

In addition, this PR is a more lightweight alternative for the closed
PR: #12243, trying to solve the same problem through implementing the
idea from #7642: "spillable flavor of objects deep into the join code,
only taking a ColumnarBatch from it when needed"

---------

Signed-off-by: sperlingxx <lovedreamf@gmail.com>
Co-authored-by: Liangcai Li <firestarmanllc@gmail.com>