Update ExprChecks for RowNumber. (#12178)
The supportedExprs.csv currently do not currently state that the
RowNumber window function does not support an `Array` order-by column.

Indeed, on running such a query, one sees that the operation falls back
to CPU:
```
scala> sql(" select row_number() over ( order by id_array ) from ... ").show
...
        !Expression <WindowSpecDefinition> windowspecdefinition(id_array#28 ASC NULLS FIRST, specifiedwindowframe(RowFrame, unboundedpreceding$(), currentrow$())) cannot run on GPU because order expression SortOrder id_array#28 ASC NULLS FIRST (ArrayType(LongType,true) is not supported)
```

This commit updates the registration of the RowNumber window fuction in
`GpuOverrides`, to allow for the generated documentation to be updated.

Note that RowNumber could well be implemented to work in this case also,
at a later date. That is out of the scope of this current change.

Signed-off-by: MithunR <mithunr@nvidia.com>