Fix Spark-4.0 build errors for sql-plugin, datagen and integration_tests module (#12300)
This contributes to https://github.com/NVIDIA/spark-rapids/issues/12062.

This PR fixes compile errors for sql-plugin, datagen and
integration_tests module.
Short description of the fixes for compile failures:
1. `AvroDataFileReader` - Spark-4.0 has dependency on Avro version
1.12.0 where setValidate function is removed. Added shim 400 and updated
the code to fix compile error.
2. Apache Spark's code still imports from
org.apache.spark.sql.SparkSession in some files. Example -
DataWritingCommandExec -
[link](https://github.com/apache/spark/blob/890ee47a9a74d48f529f3c2cffc076530e971453/sql/core/src/main/scala/org/apache/spark/sql/execution/command/commands.scala#L24).
Adhered with the same convention in this PR. In couple of files, we
import SparkSession from both spark.sql and spark.sql.classic
3. `splitFiles` signature in PartitionedFileUtil is updated in
Spark-4.0. Created separate PartitionedFileUtilShim for spark400.
4. For `datagen` module, update the package name for imports.
5. For `integration_tests` module, update the package name for import
(since sparkSession doesn't hold sqlContext anymore in Spark-4.0).
Updated it to work with all the Spark versions.

With these fixes, the error now is in `tests` module which will be
addressed in a separate PR.

---------

Signed-off-by: Niranjan Artal <nartal@nvidia.com>