Fix Spark-4.0 build errors for tests module. (#12313)
This fixes https://github.com/NVIDIA/spark-rapids/issues/12062 and fixes
https://github.com/NVIDIA/spark-rapids/issues/12309

This is the final PR to fix build errors in Spark-4.0 which was caused
to package name change issue -
https://github.com/NVIDIA/spark-rapids/issues/12062

Short description of the changes made in this PR to fix the errors:
1. Updated SparkShimServiceProvider to include a snapshot version 4.0.1
2. Added new helper methods in TrampolineConnectShims to be compatible
with Spark-4.0 and previous versions.
3. Updated the import paths to use `spark.implicits._` instead of
`spark.sqlContext.implicits._` across multiple test suites for
compatibility. sqlConext is removed from SparkSession object in
Spark-4.0.
4. Adjusted the usage of SparkSession in test suites to align with
TrampolineConnectShims.SparkSession
5. Adjusted the usage of DataFrame in test suites to align with
TrampolineConnectShims.DataFrame

---------

Signed-off-by: Niranjan Artal <nartal@nvidia.com>
Co-authored-by: Gera Shegalov <gshegalov@nvidia.com>