Run `test_write_with_stripe_size_rows` only on apache runtime and databricks runtime [databricks] (#12334)
Closes https://github.com/NVIDIA/spark-rapids/issues/12326

The root cause of this bug is that the logic in `spark_tmp_path()` uses
HDFS when it detects HDFS, while the `glob.glob` used in
`test_write_with_stripe_size_rows` only supports the local file system.
And PyArrow also requires files to be stored on the local file system.

This PR enables `test_write_with_stripe_size_rows` to run only on Apache
runtime and Databricks runtime.

---------

Signed-off-by: Yan Feng <fengyan_@mail.ustc.edu.cn>
Signed-off-by: ustcfy <fengyan_@mail.ustc.edu.cn>