Shim functions to fix binary-dedupe failures on Spark-4.0 (#12342)
This fixes https://github.com/NVIDIA/spark-rapids/issues/12332.

We fixed the build errors for Spark-4.0 in multiple PR's. The issue with
binary-dedupe failure is because of the change in signature in the below
2 functions in SparkShims.scala of unshimmed directory. The issue is due
to this import while fixing the compile errors:
https://github.com/NVIDIA/spark-rapids/commit/32dbbc509282aaa20db9ea314336fbc8df4dec14#diff-a8d9a2e2e6f805b33e5b2365b3b7f040ce7ebf810bcbe78d5721964f3ce2746cR37

Change in function signature:
For versions prior to Spark-4.0 it is:
```
  public abstract org.apache.spark.sql.SparkSession sessionFromPlan(org.apache.spark.sql.execution.SparkPlan);
  public abstract int leafNodeDefaultParallelism(org.apache.spark.sql.SparkSession); 
```
For Spark-4.0 version it is :
```
 public abstract org.apache.spark.sql.classic.SparkSession sessionFromPlan(org.apache.spark.sql.execution.SparkPlan);
 public abstract int leafNodeDefaultParallelism(org.apache.spark.sql.classic.SparkSession);
```
The fix is to remove these functions from unshimmed directory and shim
these to support different Spark versions.
Updated the supportedExprs.csv and operatorsScore.csv file for Spark4.0.

Verified with the repro case mentioned in the issue and the build
passes.

---------

Signed-off-by: Niranjan Artal <nartal@nvidia.com>