print spillable summary on bookkeep (#12338)
contributes to https://github.com/NVIDIA/spark-rapids/issues/12173

When BOOKKEEP is enabled and when it is triggered printing, it is high
likely that we encounter some bugs from oom state machine like
https://github.com/NVIDIA/spark-rapids/issues/12311. At this point, some
of the spillable handles are spillable but oom state machine might
mistakenly think there's no more memory to use.

so it's useful to print how much much memory is spillable from the
perspective from SpillFramework.

---------

Signed-off-by: Hongbin Ma (Mahone) <mahongbin@apache.org>