Fix the Pandas UDF test failures on DB14.3[databricks] (#12456)
close https://github.com/NVIDIA/spark-rapids/issues/12428

This PR is to fix all of the 3 test failures on DB14.3.

- The first one is about `RunningWindowFunctionExec` which we already
meet on DB13.3, just also mark it as xfail for 14.3.

- The second is due to the node name change from `PythonMapInArrowExec`
to `MapInArrowExec`. So update the name accordingly for the fallback
test.

- And the last is about the TaskContext type for the barrier mode. We
just relax the check for DB14.3, it is OK since it is an extra
validation and does not affect the final output.

---------

Signed-off-by: Firestarman <firestarmanllc@gmail.com>