Fix incorrect supported DataSources in generated tools generated files
Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>

Fixes #12438

This code fixes `TypeChecks.SupportedOpsForTools` to columns in the
supportedDataSource.csv
The legacy behavior assumed that the types are unknown for writeOps.
Therefore, the columns had `N/A` across most of the write formats.

This caused the Qual tool to marke all write operations as unsupported.
With the new fix, the generated CSV files should behave closer to the
generated markdown files.