Fix incorrect supported DataSources in generated tools generated files (#12481)
Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>

Fixes #12438

This code fixes `TypeChecks.SupportedOpsForTools` to columns in the
supportedDataSource.csv
The legacy behavior assumed that the types are unknown for writeOps.
Therefore, the columns had `N/A` across most of the write formats.

This caused some confusion about the level of support of each format.
Then, we had to look at the plugin code to figure out whether the format
is supported conditionally or enabled by default..etc.
With the new fix, the generated CSV files should behave closer to the
generated markdown files.

---------

Signed-off-by: Ahmed Hussein (amahussein) <a@ahussein.me>