Fix bug where rows dropped if partitioned column size is too large (#12490)
This fixes #12435 

The bug was that in a specific class when checking if another batch was
available, it totally missed that the code to add the partition columns
might have split the original input batch. This only shows up in perfile
as other classes are used in the other modes, and because perfile is not
the default it is not super likely to show up in practice.

Signed-off-by: Robert (Bobby) Evans <bobby@apache.org>