Fix the flaky ThrottlingExecutor task metrics test (#12463)
Fixes https://github.com/NVIDIA/spark-rapids/issues/12410.

The test in question incrementally increases the task wait time in a
loop, which will update the max throttle time metric in the
`ThrottlingExecutor`. #12410 reports that this might be not always the
case, which I'm not sure how it could happen except for the case where
there was some unexpected extra wait while throttling. Those unexpected
waits (or slowness) are quite common especially in the CI environment
with limited resources. I fixed the test to compute a max wait time and
use it to compare with the metric in the `ThrottlingExecutor` instead of
using the current wait time.

---------

Signed-off-by: Jihoon Son <ghoonson@gmail.com>