Accelerate ArrayDistinct (#12306)
Closes #5221 
This change will add ArrayDistinct. The implementation uses CUDF's
dropListDuplicates. We treat -0.0 and +0.0 as the same. This behavior
can differ from Spark. Spark is inconsistent in behavior, and sometime
treats -0.0 and +0.0 the same, and sometime differently. See
https://issues.apache.org/jira/browse/SPARK-51475 for more info on this.

This change is dependent on a change in cudf:
https://github.com/rapidsai/cudf/pull/18237
Without exposing a new duplicate keep option parameter, this code will
not work.

- [x] Add GpuArrayDistinct
- [x] Add test cases
- [ ] Performance benchmarking
- [x] Update documentation

---------

Signed-off-by: Warrick He <warrickh@nvidia.com>
Co-authored-by: Gera Shegalov <gshegalov@nvidia.com>