LoRE: support dump data from shuffle related nodes (#12467)
To fix: #12466 

Consider the following case:

![image](https://github.com/user-attachments/assets/017c74f0-9fc9-41cf-9b0d-67c4e2c900e6)
We want to replay the execution of `GpuShuffledSymmetricHashJoin`, thus
we need to dump the data for its child `GpuCustomShuffleReader` and
`GpuHashAggregate`.

GpuCustomShuffleReader gives out batches of `SerializedTableColumn` or
`KudoSerializedTableColumn` that current LoRE framework doesn't support
to dump and load.

To dump it this PR:

- detects if it's a SerializedTableColumn or KudoSerializedTableColumn
case
- if so convert it back to Table as a normal case(Table) to leverage
existing dump methods
- get the real data types from its plan meta and dump it along with the
data

This PR aims to add support for such case.

---------

Signed-off-by: Allen Xu <allxu@nvidia.com>