Add a physical rule to fold two-stages LocalAggregate [databricks] (#12458)
Closes #12457

This PR introduces a new Spark physical optimization rule: LocalAggregateRule, which is intended to run right before GPU optimization rules (GpuOverrides). And it had better not to be applied under Databricks runtime, so as to avoid potential conflicts due to similar optimization. Therefore, this PR injects the rule as a SparkShimRule.

An internal configure `spark.rapids.sql.foldLocalAggregate.enabled` is added for this rule, which is enabled by default.

---------

Signed-off-by: sperlingxx <lovedreamf@gmail.com>