Align the semantics of the `dynamic partitions` metric for GPU write with Spark (#12496)
close https://github.com/NVIDIA/spark-rapids/issues/11326

This PR corrects the GPU computation of the `number of dynamic part`
metric for GPU write, because it needs to match the CPU behavior and
this is an existing Spark metric.

The CPU is combining all of the reported partitions into a set on the
driver, not within each task. But the GPU was simply reporting partition
counts per task, which means we're probably over counting partitions
when multiple tasks end up writing to the same partition.

I verified it locally and it works well.

---------

Signed-off-by: Firestarman <firestarmanllc@gmail.com>