Add Delta 3.3.0 module to Maven [databricks] (#12624)
This PR adds a new Maven module for Delta 3.3.0 to Spark 3.5.x shim
profiles. It is far from working yet.

The main purpose of the PR is to refactor the code in a way that OSS
shared code compiles for 33x and works for existing shims.

The PR increases code reuse and simplifies poms w.r.t. Scala2.13 

Closes #12623

---------

Signed-off-by: Gera Shegalov <gshegalov@nvidia.com>