Supports ToPrettyString for date type in non-UTC timezone [databricks] (#12613)
closes #12537

### Analysis
Currently cast between date and string in non-UTC timezone is supported,
but ToPrettyString for date falls back.
ToPrettyString is actually a cast, so it should not fall back.

BaseExprMeta.`needTimeZoneCheck` indicate if a expr is related to
timezone.
E.g. casting date to string is not related to timezone.
ToPrettyString for date is reusing the casting date to string, but the
`needTimeZoneCheck` does not check for `ToPrettyString` expr. See:

```scala
  lazy val needTimeZoneCheck: Boolean = {
    wrapped match {
...
      case _: TimeZoneAwareExpression =>
        if (wrapped.isInstanceOf[Cast]) {
           needsTimeZone(cast.child.dataType, cast.dataType)
        } else if(PlanShims.isAnsiCast(wrapped)) {
...
        } else{
          true
        }
      case _ => false
    }
  }
```
The easy way is adding a case block to check: `case ToPrettyString`, but
ToPrettyString is not exist in Spark e.g.: 320.

### Solution
- Impl an override function for `needTimeZoneCheck` and call
`needsTimeZone(cast.child.dataType, cast.dataType)`
- BTW, a minor refactor: rename `needsTimeZone` to `castNeedsTimeZone`

Signed-off-by: Chong Gao <res_life@163.com>

---------

Signed-off-by: Chong Gao <res_life@163.com>
Co-authored-by: Chong Gao <res_life@163.com>