Allow 0 for some footer fields (#12667)
We recently ran into an situation where we got an error trying to read a
specific parquet file. It turns out that the parquet file was written by
"parquet-cpp-arrow version 13.0.0" according to the footer. It also had
1 row group with 0 rows in it. This results in us guessing wrong about
how large the footer should be after we read in the data because the
`data_page_offset` was written out as 0, and when we write it out again
we "fixed" it to a correct offset. This resulted in the size of the
footer getting larger, and us walking off the end of the buffer we had
allocated. We already work around a similar issue if we are coalescing
multiple files into a single output file, so I just made that happen all
the time and now the size is correct. Or at least large enough.

---------

Signed-off-by: Robert (Bobby) Evans <bobby@apache.org>