Add a metric to represent the max number of open writers (#12656)
This PR introduces a new metric named "max writers number" in the GPU
write exec to show the max number of writers being opened concurrently
during the write process, to help the perf analysis.

This is designed for the concurrent write, and it will always be 1 for
writer of other types if the data is not empty.

Verified by local tests.

Signed-off-by: Firestarman <firestarmanllc@gmail.com>