Use metrics about memory usage to adjust the number of concurrent tasks (#12374)
This depends on https://github.com/NVIDIA/spark-rapids-jni/pull/3056

The general idea is that we should look at how much memory each task
would need to complete successfully and adjust the concurrent level per
stage based on those numbers.

I did a bunch of benchmarks on NDS query 63 scale factor 200 to get an
idea about how to set the heuristic and currently I think that setting
future tasks in the same stage based on the 80%-tile memory usage of at
least 4 previous tasks is a nice balance.

Please note that I don't think that this is the final solution. But the
number are compelling enough that I wanted to get this in before doing a
design for the final soltuion.

---------

Signed-off-by: Robert (Bobby) Evans <bobby@apache.org>