AsyncOutputStream should copy the buffer before scheduling write task (#12681)
Fixes a bug in `AsyncOutputStream` that content in the output file can
be corrupted if the caller reuses the same buffer to the `write` calls,
which is almost always the case. Since the write task in
`AsyncOutputStream` accesses the buffer by reference in the lambda, the
buffer might have been updated already when `delegate.write()` is called
later by `ThrottlingExecutor`. To fix this, `AsyncOutputStream` should
copy the input buffer in `write`.

---------

Signed-off-by: Jihoon Son <ghoonson@gmail.com>