Use task priority for spill priority (#12766)
This depends on https://github.com/NVIDIA/spark-rapids/pull/12756 and
https://github.com/NVIDIA/spark-rapids/pull/12765

I think there is also some code cleanup that I need to do, which is why
this is still in draft.

I ran some benchmarks in local mode with the GPU memory limit set to 16
GiB, and the a kind of craze query that was designed to spill. For the
median of 3 runs this reduces the total spill by about 12%. But I am not
sure that the query is that representative, but it did show a nice
improvement.

I plan to do some follow on work for this to try and "fix" spill
priorities that are given by the user, but the improvement here was
enough that I didn't want to wait for the other code to be ready.

---------

Signed-off-by: Robert (Bobby) Evans <bobby@apache.org>
Co-authored-by: Liangcai Li <firestarmanllc@gmail.com>