Spark-4.0: Skip scala tests when ansi mode is enabled (#12813)
This PR is the last one to resolve unit test failures on Spark-4.0.
In this PR, we skip the tests when ansi mode is enabled. We need to fix
the underlying issue to enable those.

We missed this in the earlier iteration. After we merge this, we can
enable Spark-4.0 unit tests in nightly CI/CD jobs.
This fixes https://github.com/NVIDIA/spark-rapids/issues/11235

Verified by running the complete test suite:

```
mvn package -f scala2.13/pom.xml -pl tests -am -Dbuildver=400

```

---------

Signed-off-by: Niranjan Artal <nartal@nvidia.com>