Fix a potential issue that could cause a NPE when dumping kudo tables (#12862)
When enabling kudo debug mode
https://github.com/NVIDIA/spark-rapids/pull/12325, there is a potential
bug that we should not assume that the kudo `mergeOnHost()` method is
always executed in the main task thread. Otherwise, `TaskContext.get()`
will return null. So when we test async shuffle, this bug occurs and an
NPE is reported.

This pr moves the `TaskContext.get()` out of `KudoTableOperator` and
lets it happen when we build `CoalesceReadOption`, which will be on the
main thread.

---------

Signed-off-by: Haoyang Li <haoyangl@nvidia.com>