cpu mem defaults (#12680)
resolves https://github.com/NVIDIA/spark-rapids/issues/12471

Adds internal logic in the GpuDeviceManager to come up with memory
limits either based on explicitly configured spark.rapids.offHeapLimit
if set, else spark.executor.memoryOverhead if set, else base on querying
the host for available memory if possible. As a last fallback, it
defaults to a minimum of 4GB.

I tested this in different cloud environments and found all of them to
exhibit the typical docker-like mem file setup that's handled by the
code in the MemoryChecker.

Working on benchmarking with these default values versus the previous
behavior. Will post some results here soon.

---------

Signed-off-by: Zach Puller <zpuller@nvidia.com>
Co-authored-by: Alessandro Bellina <abellina@gmail.com>